modularize navigation bar, add to manager page
