+++ client/src/App.js
@@ -13,7 +13,7 @@ const App = () => {
   } else if (pageNum === 1) {
     return <EmployeePage pageUpdater={setPageNum} employeeData={employeeData}/>
   } else if (pageNum === 2) {
-    return <ManagerPage employeeData={employeeData}/>
+    return <ManagerPage pageUpdater={setPageNum} employeeData={employeeData}/>
   }
 }
