+++ client/src/components/NavigationTab.css
@@ -1,0 +1,26 @@
+
+.home-buttons{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  width: 50%;
+  padding: 5px;
+  margin: 10px;
+  background-color: rgba(255,255,255, .25);
+  border-radius: 25px;
+
+}
+
+.home-single:hover{
+  background-color: rgba(220,220,220, 1);
+}
+
+.home-single{
+  display: flex;
+  padding: 5px 15px 5px 15px;
+  border: whitesmoke 3px solid;
+  border-radius: 50px;
+  background-color: rgba(220,220,220, .5);
+  font-size: 1.5rem;
+  transition: .5s;
+}
